+++ internal/pkg/langtype/langtype.go
@@ -1,12 +1,11 @@
-package tester
+package langtype
 
 import "path/filepath"
 
 // アホ長い関数
 // 拡張子から言語を判別する
-func LanguageType(filename string) string {
-	extension := filepath.Ext(filename)
-	switch extension {
+func LangType(filename string) string {
+	switch filepath.Ext(filename) {
 	case ".c":
 		return "c"
 	case ".cc":