+++ pkg/paizaio/errors.go
@@ -1,0 +1,19 @@
+package paizaio
+
+type ClientError struct {
+	Code int
+	msg  string
+}
+
+type ServerError struct {
+	Code int
+	msg  string
+}
+
+func (e ClientError) Error() string {
+	return e.msg
+}
+
+func (e ServerError) Error() string {
+	return e.msg
+}